graphql/pyproject.toml
[tool.poetry]
name = "graphql"
version = "0.1.0"
description = ""
authors = ["Vijay Selvaraj <vijay.selvaraj@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
strawberry-graphql = {extras = ["debug-server"], version = "^0.227.2"}
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
